Split shares into multiple files
