+++ pkg/sdk/shares_enums_gen.go
@@ -1,0 +1,8 @@
+package sdk
+
+type ShareKind string
+
+const (
+	ShareKindInbound  ShareKind = "INBOUND"
+	ShareKindOutbound ShareKind = "OUTBOUND"
+)
